chore(urls): Unify SpringBoot URL handling (#3382)
