+++ hawtio-system/src/main/java/io/hawt/web/auth/LogoutServlet.java
@@ -36,7 +36,7 @@ public class LogoutServlet extends HttpServlet {
         addHeaders(response);
 
         request.logout();
-        if (AuthSessionHelpers.isSpringSecurityEnabled()) {
+        if (AuthenticationConfiguration.isSpringSecurityEnabled()) {
             AuthSessionHelpers.clear(request, authConfiguration, false);
             redirector.doRedirect(request, response, "/");
         } else {
