+++ hawtio-war-minimal/src/main/webapp/WEB-INF/web.xml
@@ -107,11 +107,11 @@
   </filter-mapping>
 
   <filter>
-    <filter-name>LoginRedirectFilter</filter-name>
-    <filter-class>io.hawt.web.auth.LoginRedirectFilter</filter-class>
+    <filter-name>ClientRouteRedirectFilter</filter-name>
+    <filter-class>io.hawt.web.auth.ClientRouteRedirectFilter</filter-class>
   </filter>
   <filter-mapping>
-    <filter-name>LoginRedirectFilter</filter-name>
+    <filter-name>ClientRouteRedirectFilter</filter-name>
     <url-pattern>/*</url-pattern>
   </filter-mapping>
